Add 16 x 16 grid and add color to each grid
Add div with class: 'grid-row' to the 'grid-container' div 16 times
using for loop. For each 'grid-row' add 16 square divs with class: 'grid' using
for loop nested within the previous for loop. 'grid-container' and
'grid-row' both have display flex and 'grid-container' has a
flex-direction of column. The end resut is a 16 x 16 grid of square divs.
Create an eventListener for each grid box. Add class: 'grid-color'
to each 'grid' when the pointer moves over it. 'grid-color' changes the
background-color to blue.