Create feature to add color style to each grid
Create function: colorGrid which takes a class to be added in the
grid where the mouse hovers. Add class to the grid only if the specific
button is clicked. Add grid-black class to the grids if the B&W button is
clicked. Remove the feature which adds a blue color to the grids when
the Grid Size button is clicked.